Add files via upload
This was not only a COP3503 problem but a 2011 ACM ICPC Southeast USA Regional Programming Contest Problem. However I solved this with a simple backtracking method, pruning off answers that could not be correct if they did not pass a MagicSum case. Again this work is not meant to be shared but to be showcased.