Add files via upload
This was another Kattis problem I solved under a specific guideline. In this problem I had to find the amount of combinations a word could take in order to fulfill certain requirements (If you read the code you'll understand) but in this I used another backtracking algorithm to prune off words that wouldn't end with a solution. Again this work is meant to be showcased not shared.